// generated/aleph-be/models/login-google-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aleph RWA Procurement API
 * Backend API for permissioned procurement pools, aggregated order tokenization, financing lifecycle, and bridge preparation on Avalanche L1.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface LoginGoogleDto {
    /**
     * Google ID token from frontend OAuth flow
     */
    'idToken': string;
}